Removal of the copybuffer
In the beginning I introduced some copy buffer with the idea that I would copy a "tile" of the image map to the copy buffer, edit the copy buffer and then copy that buffer into the output image. This step is completely redundant. and with some indexing / mapping math it's easy enough to just copy the "tile" from the tile map onto the final output image and do any editing in place there.