added robin hood and color_table refractor
After a discussion at 3DGep I was thinking if I could get rid of the unorder_map and replace it with a unordered_set and as it turns out I can! I don't need to map the color hash to the color + index, I can simply put encode the color and index in a uint_32t and give the unordered_set a custom hash and equals function that ignore the index. That way I can quickly look for a value in the set and extract the index I need from it.

Then I also realized I don't need an std::vector<int> but rather an std::vector<uint8_t> as each index has a max value of 255 and can thus fit in a uint8_t. Saving a good amount of both data and allocation time. I also combined this with a change of using a .resize over .reserve and just assigning the color index directly.