Improved index generation
I realized I was doing a lot of math to convert a 1D index to a 2D index to convert the 1D index of the actual image again. So at construction, I can calculate the proper 1D start index of the image and proper 1D end of the image. and simply iterate along that meaning no more modules, divisions and a lot less multiplications as well